More work on cleanly stopping compontents * kill started topologies * be sure to close all Curator Clients * be sure to close all ZKUtils * delete all created topics * ensure ordered shutdown in many tests
Because of Apache Storm Issue STORM-1730
[https://issues.apache.org/jira/browse/STORM-1730]  it is not possible
to shutdown the LocalServer we use for storm without leaking threads.
These threads will output error messages when they 'wake up' and try to connect
with curator - to a zk embedded server that has already been shutdown.
This issue is fixed in Storm 1.0.3, but that is not released yet

This is still a work in progress and not mergable.  There are other tests that
have issues and may need other fixes